requirements.txt
llama-index-llms-openai
llama-index-agent-openai
llama-index-core
fastapi==0.115.8
openai==1.59.3
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.31.0
supabase
pydantic>=2.0.0
# typing-extensions>=4.5.0
httpx>=0.24.0
# python-multipart>=0.0.6